+++ src/tests/password-credential.test.ts
@@ -6,6 +6,7 @@ import { PasswordCredentialService } from "../services/password-credential";
 import { mockAppObjectId, mockNewPasswordKeyId, seedMockData } from "./test-data";
 import { getTopLevelTreeItem } from "./test-utils";
 import { format } from "date-fns";
+import * as validation from "../utils/validation";
 
 // Create Jest mocks
 jest.mock("vscode");
@@ -90,8 +91,11 @@ describe("Password Credential Service Tests", () => {
 	test("Add password credential successfully", async () => {
 		// Arrange
 		jest.spyOn(passwordCredentialService, "inputDescription").mockImplementation(async () => ("Test Description"));
-		jest.spyOn(passwordCredentialService, "inputExpiryDate").mockImplementation(async (expiryDate: Date) => (format(new Date(expiryDate), "yyyy-MM-dd")));
-
+		jest.spyOn(passwordCredentialService, "inputExpiryDate").mockImplementation(async (expiryDate: Date, validation: (value: string) => string | undefined) => {
+			const expiry = format(expiryDate, "yyyy-MM-dd");
+			return validation(expiry) === undefined ? expiry : undefined;
+		});
+		
 		// Act
 		await passwordCredentialService.add(item);
 
@@ -105,6 +109,60 @@ describe("Password Credential Service Tests", () => {
 		expect(passwordItem).toBeDefined();
 		expect(passwordItem?.children?.length).toEqual(3);
 		expect(treeItem?.children?.length).toEqual(2);
+	});
+
+	test("Add password credential with incorrect date validation error", async () => {
+		// Arrange
+		const validationSpy = jest.spyOn(validation, "validatePasswordCredentialExpiryDate");
+		jest.spyOn(passwordCredentialService, "inputDescription").mockImplementation(async () => ("Test Description"));
+		jest.spyOn(passwordCredentialService, "inputExpiryDate").mockImplementation(async (_expiryDate: Date, validation: (value: string) => string | undefined) => {
+			const result = validation("NOTADATE");
+			expect(result).toBe("Expiry must be a valid date.");
+			return result;
+		});
+		
+		// Act
+		await passwordCredentialService.add(item);
+
+		// Assert
+		expect(passwordCredentialService.inputExpiryDate).toHaveBeenCalled();
+		expect(validationSpy).toHaveBeenCalled();
+	});
+
+	test("Add password credential with past date validation error", async () => {
+		// Arrange
+		const validationSpy = jest.spyOn(validation, "validatePasswordCredentialExpiryDate");
+		jest.spyOn(passwordCredentialService, "inputDescription").mockImplementation(async () => ("Test Description"));
+		jest.spyOn(passwordCredentialService, "inputExpiryDate").mockImplementation(async (_expiryDate: Date, validation: (value: string) => string | undefined) => {
+			const result = validation("2000-01-01");
+			expect(result).toBe("Expiry must be in the future.");
+			return result;
+		});
+		
+		// Act
+		await passwordCredentialService.add(item);
+
+		// Assert
+		expect(passwordCredentialService.inputExpiryDate).toHaveBeenCalled();
+		expect(validationSpy).toHaveBeenCalled();
+	});
+
+	test("Add password credential with future date validation error", async () => {
+		// Arrange
+		const validationSpy = jest.spyOn(validation, "validatePasswordCredentialExpiryDate");
+		jest.spyOn(passwordCredentialService, "inputDescription").mockImplementation(async () => ("Test Description"));
+		jest.spyOn(passwordCredentialService, "inputExpiryDate").mockImplementation(async (_expiryDate: Date, validation: (value: string) => string | undefined) => {
+			const result = validation("3000-01-01");
+			expect(result).toBe("Expiry must be less than 2 years in the future.");
+			return result;
+		});
+		
+		// Act
+		await passwordCredentialService.add(item);
+
+		// Assert
+		expect(passwordCredentialService.inputExpiryDate).toHaveBeenCalled();
+		expect(validationSpy).toHaveBeenCalled();
 	});
 
 	test("Add password credential with error", async () => {